#18 인증 추가

+++ src/main/java/com/foreveryone/knowing/controller/AdminController.java
@@ -4,8 +4,10 @@ import com.foreveryone.knowing.dto.request.InquiryRequest;
 import com.foreveryone.knowing.dto.request.ReportRequest;
 import com.foreveryone.knowing.dto.response.AdminGetResponse.InquiryResponse;
 import com.foreveryone.knowing.dto.response.AdminGetResponse.ReportResponse;
+import com.foreveryone.knowing.error.exceptions.InvalidUserTokenException;
 import com.foreveryone.knowing.service.AdminService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -18,7 +20,7 @@ public class AdminController {
 
     private final AdminService adminService;
 
-    @PostMapping("/report")
+    @PostMapping("/report") @PreAuthorize("isAuthenticated()")
     public void submitReport(@Valid @RequestBody ReportRequest reportRequest) {
         adminService.submitReport(reportRequest);
     }
@@ -29,11 +31,12 @@ public class AdminController {
     }
 
     @DeleteMapping("/{videoId}")
-    public void deleteVideo(@PathVariable Integer videoId) {
+    public void deleteVideo(@PathVariable Integer videoId, @RequestParam String secret) {
+        if (!secret.equals("secret")) throw new InvalidUserTokenException("시크릿쥬쥬");
         adminService.deleteVideo(videoId);
     }
 
-    @PostMapping("/inquiry")
+    @PostMapping("/inquiry") @PreAuthorize("isAuthenticated()")
     public void inquiry(@Valid @RequestBody InquiryRequest inquiryRequest) {
         adminService.inquiry(inquiryRequest);
     }
